requirements.txt
langchain
langchain-community
sentence-transformers
langchain_pinecone
pinecone-client
python-dotenv
streamlit
pypdf
openai
# langchain_google_community
tavily-python